sf: Silence 'Timestamp x seems implausible' logspam
Something is funky with the timestamps used in our camera's EIS routines;
they're using CLOCK_BOOTTIME instead of CLOCK_MONOTONIC, which causes
these surfaceflinger warnings to be spammed to the log. These warnings are
not fatal and there are no a/v sync issues in the camera, so silence them
in order to cut out their unneeded overhead.

Change-Id: I798d34d05139e810f9c0bdc1188e2e5cb3de8ede